+++ src/Commands/DomainControllerMakeCommand.php
@@ -3,11 +3,18 @@
 namespace Lunarstorm\LaravelDDD\Commands;
 
 use Illuminate\Routing\Console\ControllerMakeCommand;
+use Illuminate\Support\Str;
+use Lunarstorm\LaravelDDD\Commands\Concerns\CallsDomainCommands;
 use Lunarstorm\LaravelDDD\Commands\Concerns\ResolvesDomainFromInput;
+use Lunarstorm\LaravelDDD\Support\DomainResolver;
+use Lunarstorm\LaravelDDD\Support\Path;
+
+use function Laravel\Prompts\confirm;
 
 class DomainControllerMakeCommand extends ControllerMakeCommand
 {
-    use ResolvesDomainFromInput;
+    use ResolvesDomainFromInput,
+        CallsDomainCommands;
 
     protected $name = 'ddd:controller';
 
@@ -15,9 +22,13 @@ class DomainControllerMakeCommand extends ControllerMakeCommand
     {
         $modelClass = $this->parseModel($this->option('model'));
 
-        // if (! class_exists($modelClass) && confirm("A {$modelClass} model does not exist. Do you want to generate it?", default: true)) {
-        //     $this->call('make:model', ['name' => $modelClass]);
-        // }
+        if (
+            !app()->runningUnitTests()
+            && ! class_exists($modelClass)
+            && confirm("A {$modelClass} model does not exist. Do you want to generate it?", default: true)
+        ) {
+            $this->call('make:model', ['name' => $modelClass]);
+        }
 
         $replace = $this->buildFormRequestReplacements($replace, $modelClass);
 
@@ -31,6 +42,44 @@ class DomainControllerMakeCommand extends ControllerMakeCommand
             'DummyModelVariable' => lcfirst(class_basename($modelClass)),
             '{{ modelVariable }}' => lcfirst(class_basename($modelClass)),
             '{{modelVariable}}' => lcfirst(class_basename($modelClass)),
+        ]);
+    }
+
+    protected function buildFormRequestReplacements(array $replace, $modelClass)
+    {
+        [$namespace, $storeRequestClass, $updateRequestClass] = [
+            'Illuminate\\Http',
+            'Request',
+            'Request',
+        ];
+
+        if ($this->option('requests')) {
+            $namespace = $this->domain->namespaceFor('request');
+
+            [$storeRequestClass, $updateRequestClass] = $this->generateFormRequests(
+                $modelClass,
+                $storeRequestClass,
+                $updateRequestClass
+            );
+        }
+
+        $namespacedRequests = $namespace . '\\' . $storeRequestClass . ';';
+
+        if ($storeRequestClass !== $updateRequestClass) {
+            $namespacedRequests .= PHP_EOL . 'use ' . $namespace . '\\' . $updateRequestClass . ';';
+        }
+
+        return array_merge($replace, [
+            '{{ storeRequest }}' => $storeRequestClass,
+            '{{storeRequest}}' => $storeRequestClass,
+            '{{ updateRequest }}' => $updateRequestClass,
+            '{{updateRequest}}' => $updateRequestClass,
+            '{{ namespacedStoreRequest }}' => $namespace . '\\' . $storeRequestClass,
+            '{{namespacedStoreRequest}}' => $namespace . '\\' . $storeRequestClass,
+            '{{ namespacedUpdateRequest }}' => $namespace . '\\' . $updateRequestClass,
+            '{{namespacedUpdateRequest}}' => $namespace . '\\' . $updateRequestClass,
+            '{{ namespacedRequests }}' => $namespacedRequests,
+            '{{namespacedRequests}}' => $namespacedRequests,
         ]);
     }
 }