+++ src/Support/DomainAutoloader.php
@@ -9,6 +9,7 @@ use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Foundation\Application;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Str;
@@ -50,7 +51,7 @@ class DomainAutoloader
     protected static function normalizePaths($path): array
     {
         return collect($path)
-            ->filter(fn ($path) => is_dir($path))
+            ->filter(fn($path) => is_dir($path))
             ->toArray();
     }
 
@@ -99,10 +100,10 @@ class DomainAutoloader
             return Arr::wrap(Collection::times(count($classDirnameSegments), function ($index) use ($class, $classDirnameSegments) {
                 $classDirname = implode('\\', array_slice($classDirnameSegments, 0, $index));
 
-                return $classDirname.'\\Policies\\'.class_basename($class).'Policy';
+                return $classDirname . '\\Policies\\' . class_basename($class) . 'Policy';
             })->reverse()->values()->first(function ($class) {
                 return class_exists($class);
-            }) ?: [$classDirname.'\\Policies\\'.class_basename($class).'Policy']);
+            }) ?: [$classDirname . '\\Policies\\' . class_basename($class) . 'Policy']);
         });
     }
 
@@ -115,11 +116,11 @@ class DomainAutoloader
 
             $appNamespace = static::appNamespace();
 
-            $modelName = Str::startsWith($modelName, $appNamespace.'Models\\')
-                ? Str::after($modelName, $appNamespace.'Models\\')
+            $modelName = Str::startsWith($modelName, $appNamespace . 'Models\\')
+                ? Str::after($modelName, $appNamespace . 'Models\\')
                 : Str::after($modelName, $appNamespace);
 
-            return 'Database\\Factories\\'.$modelName.'Factory';
+            return 'Database\\Factories\\' . $modelName . 'Factory';
         });
     }
 
@@ -132,7 +133,7 @@ class DomainAutoloader
                 ->finish('/');
 
             $ignoredFolders = collect(config('ddd.autoload_ignore', []))
-                ->map(fn ($path) => Str::finish($path, '/'));
+                ->map(fn($path) => Str::finish($path, '/'));
 
             if ($pathAfterDomain->startsWith($ignoredFolders)) {
                 return false;
@@ -154,7 +155,9 @@ class DomainAutoloader
         }
 
         $paths = static::normalizePaths(
-            $configValue === true ? app()->basePath(DomainResolver::domainPath()) : $configValue
+            $configValue === true
+                ? app()->basePath(DomainResolver::domainPath())
+                : $configValue
         );
 
         if (empty($paths)) {