+++ tests/Generator/Model/MakeTest.php
@@ -31,7 +31,7 @@ it('can generate domain models', function ($domainPath, $domainRoot) {
 
     expect(Artisan::output())->when(
         Feature::IncludeFilepathInGeneratorCommandOutput->exists(),
-        fn ($output) => $output->toContainFilepath($relativePath),
+        fn($output) => $output->toContainFilepath($relativePath),
     );
 
     expect(file_exists($expectedModelPath))->toBeTrue();
@@ -178,4 +178,35 @@ it('skips base model creation if configured base model already exists', function
 })->with([
     ['Illuminate\Database\Eloquent\Model'],
     ['Lunarstorm\LaravelDDD\Models\DomainModel'],
+]);
+
+it('extends custom base models when applicable', function ($baseModelClass, $baseModelName) {
+    Config::set('ddd.base_model', $baseModelClass);
+
+    $domain = 'Fruits';
+    $modelName = 'Lemon';
+
+    $expectedModelPath = base_path(implode('/', [
+        config('ddd.domain_path'),
+        $domain,
+        config('ddd.namespaces.model'),
+        "{$modelName}.php",
+    ]));
+
+    if (file_exists($expectedModelPath)) {
+        unlink($expectedModelPath);
+    }
+
+    Artisan::call("ddd:model {$domain}:{$modelName}");
+
+    expect(file_exists($expectedModelPath))->toBeTrue();
+
+    expect(file_get_contents($expectedModelPath))
+        ->toContain("use {$baseModelClass};")
+        ->toContain("extends {$baseModelName}");
+})->with([
+    ['Domain\Shared\Models\BaseModel', 'BaseModel'],
+    ['Lunarstorm\LaravelDDD\Models\DomainModel', 'DomainModel'],
+    ['Illuminate\Database\Eloquent\NonExistentModel', 'NonExistentModel'],
+    ['OtherVendor\OtherPackage\Models\NonExistentModel', 'NonExistentModel'],
 ]);