+++ tests/Generator/Model/MakeWithOptionsTest.php
@@ -11,7 +11,7 @@ beforeEach(function () {
 });
 
 // it('can generate a domain model with factory', function () {
-//     $domainName = 'World';
+//     $domainName = 'Invoicing';
 //     $modelName = 'Record';
 
 //     $domain = new Domain($domainName);
@@ -53,7 +53,7 @@ beforeEach(function () {
 // });
 
 it('can generate domain model with options', function ($options, $objectType, $objectName, $expectedObjectPath) {
-    $domainName = 'World';
+    $domainName = 'Invoicing';
     $modelName = 'Record';
 
     $domain = new Domain($domainName);
@@ -90,85 +90,20 @@ it('can generate domain model with options', function ($options, $objectType, $o
         ['--factory' => true],
         'factory',
         'RecordFactory',
-        'src/Domain/World/Database/Factories/RecordFactory.php',
+        'src/Domain/Invoicing/Database/Factories/RecordFactory.php',
     ],
 
     '--seed' => [
         ['--seed' => true],
         'seeder',
         'RecordSeeder',
-        'src/Domain/World/Database/Seeders/RecordSeeder.php',
+        'src/Domain/Invoicing/Database/Seeders/RecordSeeder.php',
     ],
 
     '--policy' => [
         ['--policy' => true],
         'policy',
         'RecordPolicy',
-        'src/Domain/World/Policies/RecordPolicy.php',
-    ],
-]);
-
-it('can generate domain model with controller', function ($options, $objectType, $objectName, $expectedObjectPath) {
-    $domainName = 'World';
-    $modelName = 'Record';
-
-    $domain = new Domain($domainName);
-
-    $domainModel = $domain->model($modelName);
-
-    $expectedModelPath = base_path($domainModel->path);
-
-    if (file_exists($expectedModelPath)) {
-        unlink($expectedModelPath);
-    }
-
-    if (file_exists($expectedObjectPath)) {
-        unlink($expectedObjectPath);
-    }
-
-    $command = [
-        'ddd:model', [
-            'name' => $modelName,
-            '--domain' => $domain->dotName,
-            ...$options,
-        ],
-    ];
-
-    $this->artisan(...$command)
-        ->expectsOutputToContain(Path::normalize($domainModel->path))
-        ->assertExitCode(0);
-
-    // $output = Artisan::output();
-
-    // $outputPath = Path::normalize($domainModel->path);
-
-    // expect($output)->toContainFilepath($domainModel->path);
-
-    $path = base_path($expectedObjectPath);
-
-    expect(file_exists($path))->toBeTrue("Expecting {$objectType} to be generated at {$path}");
-
-    $contents = file_get_contents($path);
-    dump($contents);
-})->with([
-    '--controller' => [
-        ['--controller' => true],
-        'controller',
-        'RecordController',
-        'app/Http/Controllers/RecordController.php',
-    ],
-
-    '--controller --api' => [
-        ['--controller' => true, '--api' => true],
-        'controller',
-        'RecordController',
-        'app/Http/Controllers/RecordController.php',
-    ],
-
-    '--controller --requests' => [
-        ['--controller' => true, '--requests' => true],
-        'controller',
-        'RecordController',
-        'app/Http/Controllers/RecordController.php',
+        'src/Domain/Invoicing/Policies/RecordPolicy.php',
     ],
 ]);